+++ test/demo-model-test/src/main/java/com/ainrif/apiator/test/model/jaxrs/uidev/ADMIN.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2017 Ainrif <support@ainrif.com>
+ * Copyright 2014-2018 Ainrif <support@ainrif.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ainrif.apiator.test.model.jaxrs.smoke;
+package com.ainrif.apiator.test.model.jaxrs.uidev;
 
 import javax.ws.rs.HttpMethod;
 import java.lang.annotation.ElementType;
@@ -24,6 +24,6 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-@HttpMethod(HttpMethod.POST)
-public @interface CustomPOST {
+@HttpMethod("ADMIN")
+public @interface ADMIN {
 }