+++ test/jax-rs-model-test/src/main/java/com/ainrif/apiator/test/model/jaxrs/smoke/CustomHttpVerb.java
@@ -13,14 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ainrif.apiator.core.model.api
 
-enum ApiEndpointMethod {
-    GET,
-    POST,
-    PUT,
-    DELETE,
-    PATCH,
-    HEAD,
-    OPTIONS
+package com.ainrif.apiator.test.model.jaxrs.smoke;
+
+import javax.ws.rs.HttpMethod;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+@HttpMethod("CUSTOMVERB")
+public @interface CustomHttpVerb {
 }